+++ src/main/java/com/alper/worldcup/dao/SurveyResponseRepository.java
@@ -1,0 +1,7 @@
+package com.alper.worldcup.dao;
+
+import com.alper.worldcup.entity.SurveyResponse;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SurveyResponseRepository extends JpaRepository<SurveyResponse, String> {
+}