+++ src/main/java/com/alper/worldcup/entity/SurveyExtraTime.java
@@ -1,0 +1,6 @@
+package com.alper.worldcup.entity;
+
+public enum SurveyExtraTime {
+    GEREK_YOK,
+    UZATMA_EKLANSIN
+}