+++ src/main/java/com/alper/worldcup/entity/SurveyFairness.java
@@ -1,0 +1,8 @@
+package com.alper.worldcup.entity;
+
+public enum SurveyFairness {
+    KATILIYORUM,
+    KISMEN_KATILIYORUM,
+    KISMEN_KATILMIYORUM,
+    HIC_KATILMIYORUM
+}