+++ src/main/java/com/alper/worldcup/entity/SurveyScoreVisibility.java
@@ -1,0 +1,7 @@
+package com.alper.worldcup.entity;
+
+public enum SurveyScoreVisibility {
+    HERKES_GOREBILSIN,
+    SADECE_ADMIN,
+    ENCRYPTION
+}